Title:fix 27915【文件管理器】【5.1.1.33-1】【服务器】【公共】2个共享链接挂载，在左侧栏目来回切换，文件管理器崩溃 Description: RootCause:Null Solution:

+++ dde-file-manager-lib/interfaces/dfilesystemmodel.cpp
@@ -1637,10 +1637,11 @@ void DFileSystemModel::fetchMore(const QModelIndex &parent)
             d->eventLoop = Q_NULLPTR;
 
             if (code != 0) {
-                d->jobController->terminate();
-                d->jobController->quit();
-                d->jobController.clear();
-
+                if (d->jobController) { //有时候d->jobController已销毁，会导致崩溃
+                    d->jobController->terminate();
+                    d->jobController->quit();
+                    d->jobController.clear();
+                }
                 return;
             }
